[core] started working on GIF encoding
